move connections
