+++ sandbox/openfoundry_sandbox/connections/snowflake_connection.py
