--- sandbox/openfoundry_sandbox/models/connections/__init__.py
@@ -1,1 +1,0 @@
-# This file makes connections a Python package
